pyproject.toml
[tool.poetry]
name = "agentic-rag"
version = "0.1.0"
description = ""
authors = ["Tuannh"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.135.1"
uvicorn = "^0.42.0"
langgraph = "^1.1.3"
langgraph-checkpoint-postgres = "^3.0.5"
psycopg-pool = "^3.3.0"
litellm = "^1.82.4"
python-dotenv = "^1.2.2"
pydantic = "^2.12.5"
mem0ai = {extras = ["postgres"], version = "^1.0.6"}
langchain-openai = "^1.1.11"
qdrant-client = "^1.17.1"
python-multipart = "^0.0.22"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
